split tests into local and live
