+++ test/scripts/hello_world_attest_fn_call.txtar
@@ -1,0 +1,10 @@
+# [execute] call the bky-as CLI to attest a function call
+stdin fn-call.json
+exec bky-as attest-fn-call
+cp stdout result.json
+
+# [check] assert stderr and stdout expected values
+exec jq -r '.transitive_attested_function_call.claims.output | @base64d' result.json
+stdout 'Hello, World!'
+exec jq -r '.transitive_attested_function_call.logs | @base64d' result.json
+stdout 'Writing "Hello, World!" to host'
